Io branch

+++ plc_sketchbook/stage_control/d_sensor.ino
@@ -4,10 +4,10 @@
 
 void set_distance_sensor_power(bool state)
 {
-  //ADD CODE HERE
-  //P1.writeDiscrete(state, 1, 4);
+  //ADD CODE HERE, this runs at startup
+  //Set IP Address, subnet mask, default gateway? Maybe do this permanantly at a computer
 }
 
-float read_distance(){
+uint32_t read_distance(){
   return 0; //WRITE CODE HERE
 }
